separated custom player ui code into new file and deleted duplicate fullscreen button
still need to make fullscreen icon change when clicked on.
